09 21 2025
Just getting the problem set up

The database needed to be created.

I had to run and take care of that.

Good to review these things.

Had to mess with the settings of the Students table.

There is a duplicate ID number.  It is 19.  I thought I could make ID the PK (primary key).

But the fact that is had a duplicate meant that it could be a possible problem question.

or my mistake but I just copied/pasted

maybe I should check sometime.  just in case,