10 13 2025
well, I had a great time deepening my understanding of SQL.

I am right now thinking that next time I will work on the JOIN ON statement.

Speciffically as it pertains to the second part of the JOIN ON Grades.Grade
or
Students.Marks?

then yeah, more to think about and try out.